+++ examples/oxfordflowers102/diffusion/train.go
@@ -19,6 +19,7 @@ import (
 	"github.com/gomlx/gomlx/ui/gonb/margaid"
 	"github.com/gomlx/gomlx/ui/gonb/plotly"
 	stdplots "github.com/gomlx/gomlx/ui/plots"
+	"github.com/gomlx/gopjrt/dtypes"
 	"github.com/janpfeifer/must"
 	"k8s.io/klog/v2"
 	"os"
@@ -55,9 +56,10 @@ func (c *Config) AttachCheckpoint(checkpointPath string) (checkpoint *checkpoint
 		ExcludeParams(excludeParams...).
 		Done())
 	c.Checkpoint = checkpoint // Save in config.
-	fmt.Printf("\tCheckpoint: %q\n", checkpoint.Dir())
 
 	// In case the loaded checkpoint has different values, we need to update the config accordingly.
+	c.DType = must.M1(dtypes.DTypeString(
+		context.GetParamOr(c.Context, "dtype", "float32")))
 	c.ImageSize = context.GetParamOr(c.Context, "image_size", 64)
 	c.BatchSize = context.GetParamOr(c.Context, "batch_size", 64)
 	c.EvalBatchSize = context.GetParamOr(c.Context, "eval_batch_size", 128)
@@ -178,7 +180,7 @@ func TrainModel(ctx *context.Context, dataDir, checkpointPath string, paramsSet 
 		backend, ctx, config.BuildTrainingModelGraph(), customLoss,
 		optimizers.FromContext(ctx),
 		[]metrics.Interface{movingImagesLoss, movingNoiseLoss, movingMAE}, // trainMetrics
-		[]metrics.Interface{meanImagesLoss, meanMAE}) // evalMetrics
+		[]metrics.Interface{meanImagesLoss, meanMAE})                      // evalMetrics
 	if nanLogger != nil {
 		trainer.OnExecCreation(func(exec *context.Exec, _ train.GraphType) {
 			nanLogger.AttachToExec(exec)